test(core): simplify provider test setup
